==> final-notebooks/1-Clean-tdb-json.ipynb <==
# Cleaning of data to output JSON master file

## Setup

import json
import pandas as pd
import numpy as np
import os

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)

folder = './raw-data/Open-trivia-database-master/Open-trivia-database-master/en/todo/'


## Function definitions

# Function that reads input file(s) from folder and creates df
def importData(x):
    directory = os.fsencode(x)
    dataListFull = []
    
    for file in os.listdir(directory):
        filename = os.fsdecode(file)
        if filename.endswith('.json'):
            dataList = []
            with open(directory + file, encoding="utf8") as f:
                data = f.read()


            lineiterator = data.splitlines()
            for row in lineiterator:
                dataList.append(row.strip()[:-1])
            del dataList[0]
            del dataList[-1]
            dataListFull.extend(dataList)
        
    return dataListFull

# Function that converts string to dict using eval() when possible. Errors are handled in except and error text returned for log.
def dictConv1(x):
    try:
        return eval(x)
    except:
        return "ERROR-Incorrect formatting for dict conversion"

# Function that replaces double quote char with single quote within the question part of the broken JSON object.
# Returns dict if possible
# If error: returns string with potentially corrected object + info (corrected and dict success; corrected, still error; not corrected, still error)

def doubleQuoteCharConv(x):
    qStartStr = '"question":"'
    qEndStr = '", "answer"'
    stringUsed = x
    start = stringUsed.find(qStartStr) + len(qStartStr)
    end = stringUsed.find(qEndStr)
    correctionPerformed = "False"
    
    if stringUsed[start:end].find('"') != -1:
        correctionPerformed = "True"
    
    qCorrected = stringUsed[start:end].replace('"', "'")
    stringCorrected = stringUsed[:start] + qCorrected + stringUsed[end:]
    
    if correctionPerformed == 'True':
        return stringCorrected + '_correctionPerformed_' + correctionPerformed
    else:
        return '_correctionPerformed_' + correctionPerformed


# Function that saves index and best string of error strings to csv file for manual correction
def errorsToCsv(x):
    x.to_csv('./edited-data/errorsForManualCorr.csv', encoding='utf-8')

# Function that takes file with manually corrected strings, adds to df, converts to dict, updates log
# Save correction file in ./edited-data/corrected.csv
def correctionsToDf(x):
    dfTemp = pd.read_csv('./edited-data/corrected.csv', encoding='utf-8')
    dfTemp.set_index('Unnamed: 0', inplace = True)
    x['ManualCorrections'] = dfTemp['Best string']
    x.loc[x['ManualCorrections'].notnull() ,'ManualCorrections'] = x.loc[x['ManualCorrections'].notnull() ,'ManualCorrections'].apply(lambda y: dictConv1(y))

# Function that converts df to export form, exports to json
def exportJson(x):
    jsonData = json.dumps(x.tolist())
    
    with open("./edited-data/fixedJson.json", "w+") as f:
        f.write(jsonData)

## Opening data and running functions

# Run importData() function. Create pandas dataframe and add list data to "Original data str" column.
dataListFull = importData(folder)

df = pd.DataFrame(dataListFull)
df = df.rename(columns={0: "Original data str"})


# Creates 'As dict' master column in df. Runs dictConv1(), populates 'As dict' and saves outcome of function run in 'Run dictConv1'
df['Run dictConv1'] = df["Original data str"].apply(lambda x: dictConv1(x))
df['As dict'] = df['Run dictConv1'].where(df['Run dictConv1'] != 'ERROR-Incorrect formatting for dict conversion')

# Insert 'Success' in place of dict that was moved to 'As dict'
df.loc[df['Run dictConv1'] != 'ERROR-Incorrect formatting for dict conversion', 'Run dictConv1'] = 'SUCCESS'

# Change order of columns
df = df[["Original data str", "As dict", "Run dictConv1"]]

# Create 'Best string' column to store updated versions of data as string that cannot yet be converted to dict
df['Best string'] = df['Original data str']
# Create 'Row log' to store info of final row status
df['Row log'] = 'STR - No update done'
df.loc[df['Run dictConv1'] != 'ERROR-Incorrect formatting for dict conversion', 'Row log'] = 'DICT - Converted to dict on first attempt'

# Runs doubleQuoteCharConv(), adds new dicts to 'As dict'
df['Run doubleQuoteCharConv'] = df.loc[df['Run dictConv1'] != 'SUCCESS', 'Original data str'].apply(lambda x: doubleQuoteCharConv(x))
df[['doubleQuoteCharConv - str','doubleQuoteCharConv - perf']] = df['Run doubleQuoteCharConv'].str.split("_correctionPerformed_",expand=True)
df['doubleQuoteCharConv - str'].replace("", np.nan, inplace=True)
df['doubleQuoteCharConv - dict'] = df.loc[df['doubleQuoteCharConv - perf'] == 'True' ,'doubleQuoteCharConv - str'].apply(lambda x: dictConv1(x))


# 1. New dicts filtered from 'doubleQuoteCharConv - dict'
newDicts =  df.loc[df['doubleQuoteCharConv - dict'].notnull(), 'doubleQuoteCharConv - dict'].loc[df['doubleQuoteCharConv - dict'] != 'ERROR-Incorrect formatting for dict conversion']
# 2. New dicts put at right index to 'As dict'
df.loc[newDicts.index, 'As dict'] = newDicts
# 3. Update 'Best string' with rows not converted to dict
df.loc[df['doubleQuoteCharConv - perf'] == 'True', 'Best string'] = df.loc[df['doubleQuoteCharConv - perf'] == 'True', 'doubleQuoteCharConv - str']
# 4. Update logs
toDictsI = newDicts.index
newStringnoDictI = df.loc[df['doubleQuoteCharConv - dict'] == 'ERROR-Incorrect formatting for dict conversion', 'doubleQuoteCharConv - str'].index
df.loc[toDictsI, 'Row log'] = 'DICT - Corrected double quotes in q and converted to dict'
df.loc[newStringnoDictI, 'Row log'] = 'STR - Corrected double quotes in q - json format errors persist. Best string updated.'

# Select error rows and run errorsToCsv()
errorRowsforManual = df.loc[(df['Row log'] == 'STR - Corrected double quotes in q - json format errors persist. Best string updated.') | (df['Row log'] == 'STR - No update done') ,'Best string']
errorsToCsv(errorRowsforManual)

print(df.groupby('Row log').count()['Original data str'])

## Run after manual corrections done and saved to ./edited-data/corrected.csv

# Run correctionsToDf(), update 'As dict' if possible and update logs
correctionsToDf(df)

df.loc[(df['ManualCorrections'].notnull()) & (df['ManualCorrections'] != '"ERROR-Incorrect formatting for dict conversion"') , 'As dict'] = df.loc[(df['ManualCorrections'].notnull()) & (df['ManualCorrections'] != '"ERROR-Incorrect formatting for dict conversion"') , 'ManualCorrections']
df.loc[(df['ManualCorrections'].notnull()) & (df['ManualCorrections'] != '"ERROR-Incorrect formatting for dict conversion"') , 'Row log'] = 'DICT - Manually corrected'
df.loc[(df['ManualCorrections'] == '"ERROR-Incorrect formatting for dict conversion"') , 'Row log'] = 'STR - Manually corrected, still error'

print(df.groupby('Row log').count()['Original data str'])

## Export usabel JSON

exportJson(df['As dict'])
