==> final-notebooks/8-t-analysis-masterCleaning.ipynb <==
import pandas as pd
import numpy as np
import json
import matplotlib.pyplot as plt
import os

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.max_colwidth', 1000)

fig_size = plt.rcParams["figure.figsize"]
fig_size[0] = 12
fig_size[1] = 5
plt.rcParams["figure.figsize"] = fig_size

t_data = pd.read_pickle('../workproduct-files/cleaned-dataframes/t_dataMaster.pkl')
similarity = pd.read_pickle('../workproduct-files/similarityMatrix/FullSimilarityComparison.pkl')

## Set cutoff - row-pairs with similarity ratio > cutoff are processed

# Cutoff determined by estimating at what point similar rows are likely no to be true duplicates
cutoff = 0.98

similarity = similarity.loc[similarity['similarityRatio'] > cutoff]

## Select rows to drop from data master

t_data.set_index('CONS_id', inplace = True)

cols_comparison_i = ['i-category', 'i - Has category', 'i - Is all title case', 'i - Ends in punctuation', 'i - nr of cap words', 'i - source']
cols_comparison_j = ['j-category', 'j - Has category', 'j - Is all title case', 'j - Ends in punctuation', 'j - nr of cap words', 'j - source']
cols_comparison_i.extend(cols_comparison_j)
cols_comparison = cols_comparison_i
dfColsComp = pd.DataFrame(columns = cols_comparison)

similarity = similarity.append(dfColsComp)

def countCaps(x):
    i = 0
    for c in x:      
        i = i + c.isupper()
    return i

similarity['i - source'] = similarity['i-id'].apply(lambda x: t_data.loc[x, 'Source'])
similarity['j - source'] = similarity['j-id'].apply(lambda x: t_data.loc[x, 'Source'])

similarity['i-category'] = similarity['i-id'].apply(lambda x: t_data.loc[x, 'CONS_category'])
similarity['j-category'] = similarity['j-id'].apply(lambda x: t_data.loc[x, 'CONS_category'])

similarity['i - Has category'] = similarity['i-id'].apply(lambda x: not t_data.loc[x, 'CONS_category'] == 'Uncategorized')
similarity['j - Has category'] = similarity['j-id'].apply(lambda x: not t_data.loc[x, 'CONS_category'] == 'Uncategorized')

similarity['i - Is all title case'] = similarity['i-id'].apply(lambda x: t_data.loc[x, 'CONS_question'].istitle())
similarity['j - Is all title case'] = similarity['j-id'].apply(lambda x: t_data.loc[x, 'CONS_question'].istitle())

similarity['i - Ends in punctuation'] = similarity['i-id'].apply(lambda x: t_data.loc[x, 'CONS_question'][-1] in ['?', '.', '"', "'"])
similarity['j - Ends in punctuation'] = similarity['j-id'].apply(lambda x: t_data.loc[x, 'CONS_question'][-1] in ['?', '.', '"', "'"])

similarity['i - nr of cap words'] = similarity['i-id'].apply(lambda x: countCaps(t_data.loc[x, 'CONS_question']) )
similarity['j - nr of cap words'] = similarity['j-id'].apply(lambda x: countCaps(t_data.loc[x, 'CONS_question']) )

### Requirements on which to choose duplicate to drop
- 1) If one from rtg, other from tdb -> change q and category from rtg to tdb, keep tdb
- 2) Has category -> keep
- 3) Is all title case -> drop
- 4) Ends in punctuation ? . ?" -> keep
- 5) Not title case but has more capitalized words than other -> keep

def idToKeep(x): # take row from similarity, return [id to keep, question if changed, category if changed]
   
    ret = {'idToKeep': np.nan, 'idToDrop': np.nan, 'changedQ': np.nan, 'changedCat': np.nan}

       
    # Keep not rtg and substitute with rtg data
    if x['i - source'] == 'rtg' and x['j - source'] != 'rtg':
        ret['idToKeep'] = x['j-id']
        ret['changedQ'] = x['i-question']
        ret['changedCat'] = x['i-category']
        #print('3')
    elif x['j - source'] == 'rtg' and x['i - source'] != 'rtg':
        ret['idToKeep'] = x['i-id']
        ret['changedQ'] = x['j-question']
        ret['changedCat'] = x['j-category']
        #print('4')
    
    # Keep the one that has a category
    elif int(x['i - Has category']) + int(x['j - Has category']) == 1:
        if x['i - Has category'] == True:
            ret['idToKeep'] = x['i-id']
            #print('1')
        if x['j - Has category'] == True:
            ret['idToKeep'] = x['j-id']
            #print('2')
        
    # Don't keep if is all title case
    elif int(x['i - Is all title case']) + int(x['j - Is all title case']) == 1:
        if x['i - Is all title case'] == True:
            ret['idToKeep'] = x['j-id']
            #print('5')
        if x['j - Is all title case'] == True:
            ret['idToKeep'] = x['i-id']
            #print('6')
    # Keep the one that ends in punctuation
    elif int(x['i - Ends in punctuation']) + int(x['j - Ends in punctuation']) == 1:
        if x['i - Ends in punctuation'] == True:
            ret['idToKeep'] = x['j-id']
            #print('7')
        if x['j - Ends in punctuation'] == True:
            ret['idToKeep'] = x['i-id']
            #print('8')
    # Keep the one that has more capital case words
    elif x['i - nr of cap words'] > x['j - nr of cap words']:
        ret['idToKeep'] = x['i-id']
        #print('9')
    elif x['j - nr of cap words'] > x['i - nr of cap words']:
        ret['idToKeep'] = x['j-id']
        #print('10')
    # Keep i as default, change to capitalized if necessary
    else:
        if ['i - Is all title case']:
            ret['changedQ'] = x['i-question'].capitalize()
            
        ret['idToKeep'] = x['i-id']
        #print('11')
    
    # Document id to drop
    if ret['idToKeep'] == x['i-id']:
        ret['idToDrop'] = x['j-id']
    else:
        ret['idToDrop'] = x['i-id']
    
    return ret
    

%%time
masterEdits = pd.DataFrame(columns = ['idToKeep', 'idToDrop', 'changedQ', 'changedCat'])

for row in similarity.index:
    masterEdits = masterEdits.append(idToKeep(similarity.loc[row]), ignore_index = True)

## Map info to t_data and update. Save t_data as new master file

t_data = t_data.append(pd.DataFrame(columns = ['Drop row', 'qUpdate_duplRemoved', 'catUpdate_duplRemoved', 'Duplicates_dropped']))

t_data.loc[masterEdits['idToDrop'], 'Drop row'] = True

# List dropped duplicate id's in t_data['Duplicates_dropped']
t_data.loc[masterEdits['idToKeep'], 'Duplicate_removed'] = t_data.loc[masterEdits['idToKeep']].index.map(lambda x: masterEdits.loc[masterEdits['idToKeep'] == x, 'idToDrop'].tolist())


# Save updated questions and categories
t_data.loc[masterEdits['idToKeep'], 'qUpdate'] = t_data.loc[masterEdits['idToKeep']].index.map(lambda x: masterEdits.loc[masterEdits['idToKeep'] == x, 'changedQ'].tolist())
t_data.loc[masterEdits['idToKeep'], 'catUpdate'] = t_data.loc[masterEdits['idToKeep']].index.map(lambda x: masterEdits.loc[masterEdits['idToKeep'] == x, 'changedCat'].tolist())

# Clean qUpdate and catUpdate from nans, and empty lists, remove Uncategorized from catUpdate
t_data.loc[t_data['qUpdate'].notnull(), 'qUpdate'] = t_data.loc[t_data['qUpdate'].notnull(), 'qUpdate'].apply(lambda x: [y for y in x if pd.notnull(y)])
t_data.loc[t_data['catUpdate'].notnull(), 'catUpdate'] = t_data.loc[t_data['catUpdate'].notnull(), 'catUpdate'].apply(lambda x: [y for y in x if pd.notnull(y)])
t_data.loc[t_data['catUpdate'].notnull(), 'catUpdate'] = t_data.loc[t_data['catUpdate'].notnull(), 'catUpdate'].apply(lambda x: [y for y in x if y != 'Uncategorized'])

t_data.loc[t_data['qUpdate'].str.len() == 0, 'qUpdate'] = np.nan
t_data.loc[t_data['catUpdate'].str.len() == 0, 'catUpdate'] = np.nan

# Save data master with duplicate info columns
t_data.to_pickle('../workproduct-files/similarityMatrix/t_dataMaster-withDuplicateInfo.pkl')

# Update master data question and category with new data
t_data.loc[t_data['qUpdate'].notnull(), 'CONS_question'] = t_data.loc[t_data['qUpdate'].notnull(), 'qUpdate'].apply(lambda x: x[0])
t_data.loc[t_data['catUpdate'].notnull(), 'CONS_category'] = t_data.loc[t_data['catUpdate'].notnull(), 'catUpdate'].apply(lambda x: x[0])

# Reset index and rename 'CONS_id' column
t_data.reset_index(inplace = True)
t_data.rename(columns = {'index': 'CONS_id'}, inplace = True)

t_data.drop(t_data.loc[t_data['Drop row'] == True].index, inplace = True)
t_data.drop(['Drop row', 'qUpdate_duplRemoved', 'catUpdate_duplRemoved', 'Duplicates_dropped', 'qUpdate', 'catUpdate'], axis = 1, inplace = True)
t_data.reset_index(drop = True, inplace = True)

# Save new data master
t_data.to_pickle('../workproduct-files/cleaned-dataframes/t_dataMaster-duplicatesRemoved.pkl')

t_data
